refactor: extract project status and config routes
