+++ src/server.js
@@ -34,6 +34,8 @@ import { handleSettingsRoutes } from './server/routes/settings.js';
 import { handleModelRoutes } from './server/routes/models.js';
 import { handleGithubRoutes } from './server/routes/github.js';
 import { handleProjectRegistryRoutes } from './server/routes/projects.js';
+import { handleProjectStatusRoutes } from './server/routes/project/status.js';
+import { handleProjectConfigRoutes } from './server/routes/project/config.js';
 import webpush from 'web-push';
 import { config as loadDotenv } from 'dotenv';
 
@@ -3086,302 +3088,20 @@ const server = http.createServer(async (req, res) => {
 
     const subPath = pathParts.slice(subPathStart).join('/');
 
-    // GET /api/projects/:id/status
-    if (req.method === 'GET' && subPath === 'status') {
-      res.writeHead(200, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify(runner.getStatus()));
-      return;
-    }
+    const projectRouteContext = {
+      ...routeContext,
+      runner,
+      projectId,
+      subPath,
+      root: ROOT,
+      buildCustomTierMap,
+      modelTiers: MODEL_TIERS,
+      getPiModels,
+    };
 
-    // GET /api/projects/:id/logs
-    if (req.method === 'GET' && subPath === 'logs') {
-      const lines = parseInt(url.searchParams.get('lines')) || 50;
-      res.writeHead(200, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify({ logs: runner.getLogs(lines) }));
-      return;
-    }
+    if (await handleProjectStatusRoutes(req, res, url, projectRouteContext)) return;
 
-    // GET /api/projects/:id/agent-log
-    if (req.method === 'GET' && subPath === 'agent-log') {
-      const running = runner.currentAgent !== null;
-      res.writeHead(200, { 'Content-Type': 'application/json' });
-      // Resolve key label from pool
-      let keyLabel = null;
-      if (runner.currentAgentKeyId) {
-        const pool = getKeyPoolSafe();
-        keyLabel = (pool.keys || []).find(k => k.id === runner.currentAgentKeyId)?.label || null;
-      }
-      res.end(JSON.stringify({
-        running,
-        agent: runner.currentAgent,
-        model: runner.currentAgentModel,
-        keyId: runner.currentAgentKeyId || null,
-        keyLabel,
-        visibility: runner.currentAgentVisibility || { mode: 'full', issues: [] },
-        startTime: runner.currentAgentStartTime,
-        cost: runner.currentAgentCost || 0,
-        usage: runner.currentAgentUsage || null,
-        log: running ? runner.currentAgentLog : [],
-      }));
-      return;
-    }
-
-    // GET /api/projects/:id/agents
-    if (req.method === 'GET' && subPath === 'agents') {
-      res.writeHead(200, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify(runner.loadAgents()));
-      return;
-    }
-
-    // GET /api/projects/:id/agents/:name
-    if (req.method === 'GET' && subPath.startsWith('agents/') && subPath.split('/')[1]) {
-      const agentName = subPath.split('/')[1];
-      const details = runner.getAgentDetails(agentName);
-      if (!details) {
-        res.writeHead(404, { 'Content-Type': 'application/json' });
-        res.end(JSON.stringify({ error: 'Agent not found' }));
-        return;
-      }
-      res.writeHead(200, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify(details));
-      return;
-    }
-
-    // PATCH /api/projects/:id/agents/:name - Update agent settings
-    if (req.method === 'PATCH' && subPath.startsWith('agents/') && subPath.split('/')[1]) {
-      const agentName = subPath.split('/')[1];
-      let body = '';
-      req.on('data', chunk => body += chunk);
-      req.on('end', () => {
-        try {
-          const { model } = JSON.parse(body);
-          if (!model && model !== '') throw new Error('Missing model');
-          
-          // Check if this is a manager or worker
-          const managersDir = path.join(ROOT, 'agent', 'managers');
-          const workersDir = runner.workerSkillsDir;
-          const isManager = fs.existsSync(path.join(managersDir, `${agentName}.md`));
-          const isWorker = fs.existsSync(path.join(workersDir, `${agentName}.md`));
-          
-          if (!isManager && !isWorker) {
-            res.writeHead(404, { 'Content-Type': 'application/json' });
-            res.end(JSON.stringify({ error: 'Agent not found' }));
-            return;
-          }
-          
-          if (isManager) {
-            // Manager settings go in project config.yaml (not the skill file)
-            const config = runner.loadConfig();
-            if (!config.managers) config.managers = {};
-            if (!config.managers[agentName]) config.managers[agentName] = {};
-            if (model) {
-              config.managers[agentName].model = model;
-            } else {
-              // Empty string = clear override (inherit from skill file)
-              delete config.managers[agentName].model;
-              if (Object.keys(config.managers[agentName]).length === 0) {
-                delete config.managers[agentName];
-              }
-            }
-            fs.writeFileSync(runner.configPath, yaml.dump(config, { lineWidth: -1 }));
-          } else {
-            // Worker settings go in the skill file frontmatter
-            const skillPath = path.join(workersDir, `${agentName}.md`);
-            let content = fs.readFileSync(skillPath, 'utf-8');
-            if (content.startsWith('---')) {
-              if (model) {
-                content = content.replace(/^(---[\s\S]*?)model:\s*.+$/m, `$1model: ${model}`);
-                if (!content.match(/^model:/m)) {
-                  content = content.replace(/^---\n/, `---\nmodel: ${model}\n`);
-                }
-              }
-            } else {
-              content = `---\n${model ? `model: ${model}\n` : ''}---\n${content}`;
-            }
-            fs.writeFileSync(skillPath, content);
-          }
-          
-          res.writeHead(200, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ success: true }));
-        } catch (e) {
-          res.writeHead(400, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ error: e.message }));
-        }
-      });
-      return;
-    }
-
-    // GET /api/projects/:id/config
-    if (req.method === 'GET' && subPath === 'config') {
-      const raw = fs.existsSync(runner.configPath) ? fs.readFileSync(runner.configPath, 'utf-8') : '';
-      const config = runner.loadConfig();
-      const projectToken = config.setupToken;
-      const hasProjectToken = !!projectToken;
-      const safeConfig = { ...config };
-      delete safeConfig.setupToken;
-      // Include key pool and selection info
-      const keyPool = getKeyPoolSafe();
-      const keySelection = config.keySelection || null;
-
-      // Determine effective provider from key selection
-      let detectedKey = null;
-      if (keySelection?.keyId) {
-        detectedKey = keyPool.keys.find(k => k.id === keySelection.keyId) || null;
-      }
-      if (!detectedKey) {
-        detectedKey = keyPool.keys.find(k => k.enabled) || null;
-      }
-      const detectedProvider = detectedKey?.provider || 'anthropic';
-      const detectedTiers = detectedProvider === 'custom' && detectedKey?.customConfig
-        ? buildCustomTierMap(detectedKey.customConfig)
-        : (MODEL_TIERS[detectedProvider] || {});
-
-      // Build available models list per provider from pi-ai
-      // Only show recent/relevant models, not the full historical catalog
-      const EFFORT_LEVELS = ['medium', 'high', 'xhigh'];
-      const ALLOWED_MODELS = {
-        anthropic: /^(claude-opus-4-7|claude-sonnet-4-6)$|^claude-haiku-4-5-/,
-        openai: /^(gpt-5\.5|o[34])/,
-        'openai-codex': /^(gpt-5\.5)/,
-        google: /^gemini-[23]/,
-        minimax: /MiniMax/,
-      };
-      const availableModels = {};
-      for (const provider of Object.keys(MODEL_TIERS)) {
-        try {
-          const models = getPiModels(provider);
-          const filter = ALLOWED_MODELS[provider];
-          const entries = [];
-          for (const m of models) {
-            if (filter && !filter.test(m.id)) continue;
-            if (m.id.includes('latest')) continue; // skip aliases, use exact versions
-            if (m.reasoning) {
-              for (const effort of EFFORT_LEVELS) {
-                entries.push({ id: `${m.id}@${effort}`, name: `${m.name} (${effort})` });
-              }
-            } else {
-              entries.push({ id: m.id, name: m.name });
-            }
-          }
-          availableModels[provider] = entries;
-        } catch { availableModels[provider] = []; }
-      }
-      availableModels.custom = [];
-
-      res.writeHead(200, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify({
-        config: safeConfig, raw, hasProjectToken,
-        projectTokenPreview: projectToken ? maskToken(projectToken) : null,
-        provider: detectedProvider,
-        tiers: detectedTiers,
-        allTiers: detectedProvider === 'custom' ? { ...MODEL_TIERS, custom: detectedTiers } : MODEL_TIERS,
-        availableModels,
-        keyPool,
-        keySelection,
-        allowCustomProvider: ALLOW_CUSTOM_PROVIDER,
-      }));
-      return;
-    }
-
-    // POST /api/projects/:id/token — set per-project key selection or direct token
-    if (req.method === 'POST' && subPath === 'token') {
-      if (!requireWrite(req, res)) return;
-      let body = '';
-      req.on('data', c => body += c);
-      req.on('end', () => {
-        try {
-          const { keyId, fallback, token, provider: explicitProvider, customConfig } = JSON.parse(body);
-          const configPath = runner.configPath;
-          const existing = fs.existsSync(configPath) ? yaml.load(fs.readFileSync(configPath, 'utf-8')) || {} : {};
-
-          if (keyId !== undefined) {
-            // New key pool selection mode
-            delete existing.setupToken;
-            delete existing.setupTokenProvider;
-            if (keyId) {
-              existing.keySelection = { keyId, fallback: fallback !== false };
-            } else {
-              // Clear selection (use global default)
-              delete existing.keySelection;
-            }
-          } else {
-            // Legacy mode: raw token
-            if (token) {
-              existing.setupToken = token;
-              if (explicitProvider) existing.setupTokenProvider = explicitProvider;
-              // Also add to global pool
-              const entry = addKey({ label: `${explicitProvider || 'API'} (from ${projectId})`, token, provider: explicitProvider, customConfig });
-              existing.keySelection = { keyId: entry.id, fallback: true };
-            } else {
-              delete existing.setupToken;
-              delete existing.setupTokenProvider;
-              delete existing.keySelection;
-            }
-          }
-
-          fs.writeFileSync(configPath, yaml.dump(existing));
-          res.writeHead(200, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({
-            success: true,
-            hasProjectToken: !!(existing.setupToken || existing.keySelection?.keyId),
-            keySelection: existing.keySelection || null,
-          }));
-        } catch (e) {
-          res.writeHead(400, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ error: e.message }));
-        }
-      });
-      return;
-    }
-
-    // POST /api/projects/:id/config
-    if (req.method === 'POST' && subPath === 'config') {
-      if (!requireWrite(req, res)) return;
-      let body = '';
-      req.on('data', c => body += c);
-      req.on('end', () => {
-        try {
-          const { content } = JSON.parse(body);
-          runner.saveConfig(content);
-          res.writeHead(200, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ success: true }));
-        } catch (e) {
-          res.writeHead(400, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ error: e.message }));
-        }
-      });
-      return;
-    }
-
-    // POST /api/projects/:id/models — update project-level model overrides
-    if (req.method === 'POST' && subPath === 'models') {
-      if (!requireWrite(req, res)) return;
-      let body = '';
-      req.on('data', c => body += c);
-      req.on('end', () => {
-        try {
-          const { models } = JSON.parse(body);
-          // Read existing config, merge models, save
-          const config = runner.loadConfig();
-          if (models && (models.high || models.mid || models.low)) {
-            config.models = {};
-            if (models.high) config.models.high = models.high;
-            if (models.mid) config.models.mid = models.mid;
-            if (models.low) config.models.low = models.low;
-          } else {
-            delete config.models;
-          }
-          const newYaml = yaml.dump(config, { lineWidth: -1 });
-          runner.saveConfig(newYaml);
-          res.writeHead(200, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ success: true, models: config.models || null }));
-        } catch (e) {
-          res.writeHead(400, { 'Content-Type': 'application/json' });
-          res.end(JSON.stringify({ error: e.message }));
-        }
-      });
-      return;
-    }
+    if (await handleProjectConfigRoutes(req, res, url, projectRouteContext)) return;
 
     // GET /api/projects/:id/comments
     if (req.method === 'GET' && subPath === 'comments') {
